Reorganize the plugin config API objects
